API for Tags

+++ lib/core/database/services/tags/tags_service.dart
@@ -1,6 +1,9 @@
 import 'package:drift/drift.dart';
+import 'package:parsa/core/api/post_methods/post_user_account.dart';
+import 'package:parsa/core/api/post_methods/post_user_tag_service.dart';
 import 'package:parsa/core/database/app_db.dart';
 import 'package:parsa/core/models/tags/tag.dart';
+import 'package:parsa/core/services/auth/auth0_class.dart';
 
 class TagService {
   final AppDB db;
@@ -8,16 +11,74 @@ class TagService {
   TagService._(this.db);
   static final TagService instance = TagService._(AppDB.instance);
 
-  Future<int> insertTag(TagInDB tag) {
-    return db.into(db.tags).insert(tag);
+  Future<int> insertTag(TagInDB tag) async {
+    try {
+      // Retrieve the access token from your authentication service
+
+      final auth0 = getAuth0Instance();
+
+      // Retrieve the access token from the Auth0 instance
+      final credentials = await auth0.credentialsManager.credentials();
+      // Post the account to the API
+      bool isPosted = await PostUserTagService.postUserTag(
+          tag, credentials.accessToken, 'POST');
+
+      if (!isPosted) {
+        throw Exception('Failed to post account to the API.');
+      } else {
+        return await db
+            .into(db.tags)
+            .insert(tag, mode: InsertMode.insertOrReplace);
+      }
+    } catch (e) {
+      print('Error inserting account: $e');
+      rethrow; // Propagate the error to be handled upstream if needed
+    }
   }
 
-  Future<bool> updateTag(TagInDB tag) {
-    return db.update(db.tags).replace(tag);
+  Future<bool> updateTag(TagInDB tag) async {
+    try {
+      // Retrieve the access token from your authentication service
+
+      final auth0 = getAuth0Instance();
+
+      // Retrieve the access token from the Auth0 instance
+      final credentials = await auth0.credentialsManager.credentials();
+      // Post the account to the API
+      bool isPut = await PostUserTagService.postUserTag(
+          tag, credentials.accessToken, 'PUT');
+
+      if (!isPut) {
+        throw Exception('Failed to post account to the API.');
+      } else {
+        return db.update(db.tags).replace(tag);
+      }
+    } catch (e) {
+      print('Error updating account: $e');
+      rethrow; // Propagate the error to be handled upstream if needed
+    }
   }
 
-  Future<int> deleteTag(String tagId) {
-    return (db.delete(db.tags)..where((tbl) => tbl.id.equals(tagId))).go();
+  Future<int> deleteTag(String tagId) async {
+    try {
+      // Added missing try block
+      final auth0 = getAuth0Instance();
+
+      // Retrieve the access token from the Auth0 instance
+      final credentials = await auth0.credentialsManager.credentials();
+      // Post the account to the API
+      bool isPut = await PostUserTagService.deleteUserTag(
+          tagId, credentials.accessToken);
+
+      if (!isPut) {
+        throw Exception('Failed to post account to the API.');
+      } else {
+        return (db.delete(db.tags)..where((tbl) => tbl.id.equals(tagId))).go();
+      } // Added missing closing bracket
+    } catch (e) {
+      print('Error updating account: $e');
+      rethrow; // Propagate the error to be handled upstream if needed
+    }
   }
 
   Stream<List<Tag>> getTags({
